requirements.txt
asgiref==3.8.1
Django==4.2.16
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.1
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
typing-extensions==4.12.2
uritemplate==4.1.1
